chore: add 모듈화
