+++ Ino/week1/TODO/app/screen/ToDoList.tsx
@@ -1,8 +1,7 @@
-import React, {useState} from 'react';
-import {View, FlatList, StyleSheet, TextInput, Button} from 'react-native';
+import React from 'react';
+import {View, FlatList, StyleSheet} from 'react-native';
 import {useToDoStore} from '../stores/ToDoStore';
 import ToDoItem from '../component/ToDoItem';
-import {Text} from 'react-native-svg';
 
 export function ListScreen() {
   const {toDoList} = useToDoStore();
@@ -16,31 +15,6 @@ export function ListScreen() {
           <ToDoItem id={item.id} title={item.title} status={item.status} />
         )}
       />
-    </View>
-  );
-}
-
-export function AddScreen() {
-  const [title, setTitle] = useState<string>(''); // 새로운 ToDo의 제목 상태
-  const {addToDo} = useToDoStore(); // Zustand를 사용하여 ToDo를 추가하는 함수 가져오기
-
-  // 새로운 ToDo를 추가하는 함수
-  const handleAddToDo = () => {
-    if (title.trim() !== '') {
-      addToDo(title); // Zustand를 사용하여 ToDo를 추가
-      setTitle(''); // 입력 필드 비우기
-    }
-  };
-
-  return (
-    <View style={styles.container}>
-      <TextInput
-        style={styles.input}
-        placeholder="할일을 입력하세요"
-        value={title}
-        onChangeText={setTitle}
-      />
-      <Button title="추가" onPress={handleAddToDo} />
     </View>
   );
 }
